refactor and heartbeating resume
