+++ posthog/temporal/data_imports/pipelines/stripe/settings.py
